New updates to get_vince.py
